server/requirements.txt
openenv-core
pydantic
requests
openai
